requirements.txt
fastapi
uvicorn
jinja2
openai
# Hugging Face dependencies for Whisper-large-v3-turbo
transformers>=4.36.0
datasets
accelerate
safetensors
sentencepiece
# Additional dependencies for audio processing
Pillow
# Remove original whisper as we're using the Hugging Face version
# whisper
numpy
scipy
python-multipart
markdown
pymupdf
ffmpeg-python
pydub
python-dotenv
# OlmOCR dependencies
git+https://github.com/allenai/olmocr.git